moved strategy-components to subdirectory
