+++ probdiffeq/solvers/strategies/components/__init__.py
@@ -1,0 +1,1 @@
+"""State-space model components."""
