+++ probdiffeq/solvers/strategies/components/cubature.py
