+++ tests/test_ivpsolve/test_solution_object.py
@@ -6,7 +6,8 @@ from probdiffeq import adaptive, ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated
-from probdiffeq.solvers.strategies import correction, filters, priors
+from probdiffeq.solvers.strategies import filters
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
