+++ tests/test_solvers/test_calibrated/test_dynamic_calibration_on_exponential.py
@@ -9,7 +9,8 @@ import jax.numpy as jnp
 from probdiffeq import ivpsolve
 from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated
-from probdiffeq.solvers.strategies import correction, filters, priors
+from probdiffeq.solvers.strategies import filters
+from probdiffeq.solvers.strategies.components import correction, priors
 from tests.setup import setup
