+++ tests/test_solvers/test_calibrated/test_mle_calibration_vs_calibrationfree.py
@@ -10,7 +10,8 @@ from probdiffeq import adaptive, ivpsolve, timestep
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated, solution, uncalibrated
-from probdiffeq.solvers.strategies import correction, filters, fixedpoint, priors
+from probdiffeq.solvers.strategies import filters, fixedpoint
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
