+++ tests/test_solvers/test_misc/test_warnings_for_wrong_strategies.py
@@ -5,7 +5,8 @@ from probdiffeq import adaptive, ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import uncalibrated
-from probdiffeq.solvers.strategies import correction, fixedpoint, priors, smoothers
+from probdiffeq.solvers.strategies import fixedpoint, smoothers
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
