+++ tests/test_solvers/test_solution/test_log_marginal_likelihood_terminal_values.py
@@ -5,13 +5,8 @@ from probdiffeq import adaptive, ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import solution, uncalibrated
-from probdiffeq.solvers.strategies import (
-    correction,
-    filters,
-    fixedpoint,
-    priors,
-    smoothers,
-)
+from probdiffeq.solvers.strategies import filters, fixedpoint, smoothers
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
