+++ tests/test_solvers/test_strategies/test_corrections/test_rmse_of_correction.py
@@ -7,7 +7,8 @@ from probdiffeq import adaptive, ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import calibrated
-from probdiffeq.solvers.strategies import correction, cubature, filters, priors
+from probdiffeq.solvers.strategies import filters
+from probdiffeq.solvers.strategies.components import correction, cubature, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
