+++ tests/test_solvers/test_strategies/test_filter_vs_smoother_rmse.py
@@ -7,7 +7,8 @@ from probdiffeq import ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import uncalibrated
-from probdiffeq.solvers.strategies import correction, filters, priors, smoothers
+from probdiffeq.solvers.strategies import filters, smoothers
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
