+++ tests/test_solvers/test_strategies/test_smoother_vs_fixedpoint_equivalence.py
@@ -9,7 +9,8 @@ from probdiffeq import adaptive, ivpsolve
 from probdiffeq.backend import testing
 from probdiffeq.impl import impl
 from probdiffeq.solvers import solution, uncalibrated
-from probdiffeq.solvers.strategies import correction, fixedpoint, priors, smoothers
+from probdiffeq.solvers.strategies import fixedpoint, smoothers
+from probdiffeq.solvers.strategies.components import correction, priors
 from probdiffeq.solvers.taylor import autodiff
 from tests.setup import setup
