+++ tests/test_util/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for utilities."""
